fix condition

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -9,15 +9,15 @@
 int main(void)
 {
 	int i;
-	unsigned long int prev1 = 1, prev2 = 2, s = 0;
+	unsigned long long int prev1 = 1, prev2 = 2, s = 0;
 
-	printf("%lu, %lu, ", prev1, prev2);
+	printf("%llu, %llu, ", prev1, prev2);
 	for (i = 3; i <= 98; i++)
 	{
 		s = prev1 + prev2;
 		prev1 = prev2;
 		prev2 = s;
-		printf("%lu", s);
+		printf("%llu", s);
 		if (i != 98)
 			printf(", ");
 	}
